+++ resources/views/friend/list.blade.php
@@ -1,19 +1,23 @@
 @extends('layouts.classic')
 
-@section('title', 'Friends')
+@section('title', $owner->is(auth()->user()) ? 'Friends' : $owner->name . "'s friends")
 
 @section('content')
     <div class="dparts" id="friend_list">
-        <h2 class="partsHeading">Friends</h2>
+        <h2 class="partsHeading">
+            {{ $owner->is(auth()->user()) ? 'Friends' : $owner->name . "'s friends" }}
+        </h2>
         <div class="parts">
             @if ($friends->isEmpty())
-                <p>You have no friends yet.</p>
+                <p>No friends to show.</p>
             @else
                 <ul class="friendList">
                     @foreach ($friends as $friend)
                         <li>
                             <span class="memberName">{{ $friend->name }}</span>
-                            <a href="{{ route('friend.unlink.show', $friend) }}">Unfriend</a>
+                            @if ($owner->is(auth()->user()))
+                                <a href="{{ route('friend.unlink.show', $friend) }}">Unfriend</a>
+                            @endif
                         </li>
                     @endforeach
                 </ul>
